bugfix@BC-1500: Audit Issue#7 Prefer explicit unsigned integer sizes

+++ contracts/compliance/WalletManager.sol
@@ -78,7 +78,7 @@ contract WalletManager is IDSWalletManager, WalletManagerDataStore, BaseDSContra
    */
     function addIssuerWallets(address[] calldata _wallets) public override onlyIssuerOrAbove returns (bool) {
         require(_wallets.length <= 30, "Exceeded the maximum number of wallets");
-        for (uint i = 0; i < _wallets.length; i++) {
+        for (uint8 i = 0; i < _wallets.length; i++) {
             addIssuerWallet(_wallets[i]);
         }
         return true;
@@ -100,7 +100,7 @@ contract WalletManager is IDSWalletManager, WalletManagerDataStore, BaseDSContra
    */
     function addPlatformWallets(address[] calldata _wallets) public override onlyIssuerOrAbove returns (bool) {
         require(_wallets.length <= 30, "Exceeded the maximum number of wallets");
-        for (uint i = 0; i < _wallets.length; i++) {
+        for (uint8 i = 0; i < _wallets.length; i++) {
             addPlatformWallet(_wallets[i]);
         }
         return true;
